+++ bsdploy/fabutils.py
@@ -1,4 +1,4 @@
-from fabric.api import env, local, run, sudo
+from fabric.api import env, local, run, sudo, task
 from fabric.contrib.project import rsync_project as _rsync_project
 from ploy.common import shjoin
 
@@ -19,6 +19,7 @@ def rsync_project(*args, **kwargs):
         _rsync_project(*args, **kwargs)
 
 
+@task
 def rsync(*args, **kwargs):
     """ wrapper around the rsync command.
 
@@ -52,12 +53,28 @@ def rsync(*args, **kwargs):
     return local(cmd, **kwargs)
 
 
+@task
 def service(service=None, action='status'):
     if service is None:
         exit("You must provide a service name")
     sudo('service %s %s' % (service, action), warn_only=True)
 
 
-def freebsd_update():
-    run('freebsd-update cron')
-    run('freebsd-update install')
+
+@task
+def pkg_upgrade():
+    run('pkg-static update')
+    run('pkg-static install -U pkg')
+    run('pkg update')
+    run('pkg upgrade')
+    print("Done.")
+
+
+@task
+def update_flavour_pkg():
+    """upgrade the pkg tool of the base flavour (so that newly created jails have the latest version)"""
+    base_cmd = 'pkg-static -r /usr/jails/flavours/bsdploy_base'
+    run('%s update' % base_cmd)
+    run('%s install -U pkg' % base_cmd)
+    run('%s update' % base_cmd)
+    print("Done.")